+++ Jcd.BitManipulation/ReadBitsExtensions.cs
@@ -11,7 +11,7 @@ namespace Jcd.BitManipulation
         /// <param name="offset">the bit offset to read from</param>
         /// <param name="size">The total number of bits to extract</param>
         /// <returns>The numeric value stored at that bit location</returns>
-        public static sbyte ReadBits(this sbyte self, byte offset, byte size)
+        public static sbyte ReadBits(this sbyte self, int offset, int size)
             => self.ReadBits(offset, BitMask.CreateRange(offset, size));
 
         /// <summary>
@@ -21,7 +21,7 @@ namespace Jcd.BitManipulation
         /// <param name="offset">the bit offset to read from</param>
         /// <param name="size">The total number of bits to extract</param>
         /// <returns>The numeric value stored at that bit location</returns>
-        public static short ReadBits(this short self, byte offset, byte size)
+        public static short ReadBits(this short self, int offset, int size)
             => self.ReadBits(offset, BitMask.CreateRange(offset, size));
 
         /// <summary>
@@ -31,7 +31,7 @@ namespace Jcd.BitManipulation
         /// <param name="offset">the bit offset to read from</param>
         /// <param name="size">The total number of bits to extract</param>
         /// <returns>The numeric value stored at that bit location</returns>
-        public static int ReadBits(this int self, byte offset, byte size)
+        public static int ReadBits(this int self, int offset, int size)
             => self.ReadBits(offset, BitMask.CreateRange(offset, size));
 
         /// <summary>
@@ -41,7 +41,7 @@ namespace Jcd.BitManipulation
         /// <param name="offset">the bit offset to read from</param>
         /// <param name="size">The total number of bits to extract</param>
         /// <returns>The numeric value stored at that bit location</returns>
-        public static long ReadBits(this long self, byte offset, byte size)
+        public static long ReadBits(this long self, int offset, int size)
             => self.ReadBits(offset, BitMask.CreateRange(offset, size));
 
         /// <summary>
@@ -51,7 +51,7 @@ namespace Jcd.BitManipulation
         /// <param name="offset">the bit offset to read from</param>
         /// <param name="size">The total number of bits to extract</param>
         /// <returns>The numeric value stored at that bit location</returns>
-        public static byte ReadBits(this byte self, byte offset, byte size)
+        public static byte ReadBits(this byte self, int offset, int size)
             => self.ReadBits(offset, BitMask.CreateRange(offset, size));
        
         /// <summary>
@@ -61,7 +61,7 @@ namespace Jcd.BitManipulation
         /// <param name="offset">the bit offset to read from</param>
         /// <param name="size">The total number of bits to extract</param>
         /// <returns>The numeric value stored at that bit location</returns>
-        public static ushort ReadBits(this ushort self, byte offset, byte size)
+        public static ushort ReadBits(this ushort self, int offset, int size)
             => self.ReadBits(offset, BitMask.CreateRange(offset, size));
         
         /// <summary>
@@ -71,7 +71,7 @@ namespace Jcd.BitManipulation
         /// <param name="offset">the bit offset to read from</param>
         /// <param name="size">The total number of bits to extract</param>
         /// <returns>The numeric value stored at that bit location</returns>
-        public static uint ReadBits(this uint self, byte offset, byte size)
+        public static uint ReadBits(this uint self, int offset, int size)
             => self.ReadBits(offset, BitMask.CreateRange(offset, size));
         
         /// <summary>
@@ -81,7 +81,7 @@ namespace Jcd.BitManipulation
         /// <param name="offset">the bit offset to read from</param>
         /// <param name="size">The total number of bits to extract</param>
         /// <returns>The numeric value stored at that bit location</returns>
-        public static ulong ReadBits(this ulong self, byte offset, byte size)
+        public static ulong ReadBits(this ulong self, int offset, int size)
             => self.ReadBits(offset, BitMask.CreateRange(offset, size));
 
         #endregion
@@ -97,7 +97,7 @@ namespace Jcd.BitManipulation
         /// Zeroed bits in the mask will always extract 0 from the source.
         /// </param>
         /// <returns>The right shifted value extracted from the value</returns>
-        public static ulong ReadBits(this ulong self, byte offset, BitMask mask)
+        public static ulong ReadBits(this ulong self, int offset, BitMask mask)
         {
             return ((self & mask) >> offset);
         }
@@ -111,7 +111,7 @@ namespace Jcd.BitManipulation
         /// Zeroed bits in the mask will always extract 0 from the source.
         /// </param>
         /// <returns>The right shifted value extracted from the value</returns>
-        public static long ReadBits(this long self, byte offset, BitMask mask)
+        public static long ReadBits(this long self, int offset, BitMask mask)
         {
             return (long)(((ulong)self & mask) >> offset);
         }
@@ -126,7 +126,7 @@ namespace Jcd.BitManipulation
         /// Zeroed bits in the mask will always extract 0 from the source.
         /// </param>
         /// <returns>The right shifted value extracted from the value</returns>
-        public static uint ReadBits(this uint self, byte offset, BitMask mask)
+        public static uint ReadBits(this uint self, int offset, BitMask mask)
         {
             return ((self & mask) >> offset);
         }
@@ -140,7 +140,7 @@ namespace Jcd.BitManipulation
         /// Zeroed bits in the mask will always extract 0 from the source.
         /// </param>
         /// <returns>The right shifted value extracted from the value</returns>
-        public static int ReadBits(this int self, byte offset, BitMask mask)
+        public static int ReadBits(this int self, int offset, BitMask mask)
         {
             return (int)(((uint)self & mask) >> offset); // coerce to unsigned first to prevent issues with sign bit.
         }
@@ -154,7 +154,7 @@ namespace Jcd.BitManipulation
         /// Zeroed bits in the mask will always extract 0 from the source.
         /// </param>
         /// <returns>The right shifted value extracted from the value</returns>
-        public static ushort ReadBits(this ushort self, byte offset, BitMask mask)
+        public static ushort ReadBits(this ushort self, int offset, BitMask mask)
         {
             return (ushort)((self & mask) >> offset);
         }
@@ -168,7 +168,7 @@ namespace Jcd.BitManipulation
         /// Zeroed bits in the mask will always extract 0 from the source.
         /// </param>
         /// <returns>The right shifted value extracted from the value</returns>
-        public static short ReadBits(this short self, byte offset, BitMask mask)
+        public static short ReadBits(this short self, int offset, BitMask mask)
         {
             return (short)(((ushort)self & mask) >> offset);
         }
@@ -182,7 +182,7 @@ namespace Jcd.BitManipulation
         /// Zeroed bits in the mask will always extract 0 from the source.
         /// </param>
         /// <returns>The right shifted value extracted from the value</returns>
-        public static byte ReadBits(this byte self, byte offset, BitMask mask)
+        public static byte ReadBits(this byte self, int offset, BitMask mask)
         {
             return (byte)((self & mask) >> offset);
         }
@@ -196,7 +196,7 @@ namespace Jcd.BitManipulation
         /// Zeroed bits in the mask will always extract 0 from the source.
         /// </param>
         /// <returns>The right shifted value extracted from the value</returns>
-        public static sbyte ReadBits(this sbyte self, byte offset, BitMask mask)
+        public static sbyte ReadBits(this sbyte self, int offset, BitMask mask)
         {
             return (sbyte)(((byte)self & mask) >> offset);
         }
@@ -211,8 +211,8 @@ namespace Jcd.BitManipulation
         /// <param name="self">the bits to read from</param>
         /// <param name="offset">the offset to read</param>
         /// <returns>the bit (a bool) at the specified offset </returns>
-        public static bool ReadBit(this ulong self, byte offset)
-            => self.ReadBits(offset,(byte)1) > 0;
+        public static bool ReadBit(this ulong self, int offset)
+            => self.ReadBits(offset,1) > 0;
 
         /// <summary>
         /// Read a single bit from the specified offset.
@@ -220,8 +220,8 @@ namespace Jcd.BitManipulation
         /// <param name="self">the bits to read from</param>
         /// <param name="offset">the offset to read</param>
         /// <returns>the bit (a bool) at the specified offset </returns>
-        public static bool ReadBit(this long self, byte offset)
-            => self.ReadBits(offset,(byte)1) > 0;
+        public static bool ReadBit(this long self, int offset)
+            => self.ReadBits(offset,1) > 0;
         
         /// <summary>
         /// Read a single bit from the specified offset.
@@ -229,8 +229,8 @@ namespace Jcd.BitManipulation
         /// <param name="self">the bits to read from</param>
         /// <param name="offset">the offset to read</param>
         /// <returns>the bit (a bool) at the specified offset </returns>
-        public static bool ReadBit(this uint self, byte offset)
-            => self.ReadBits(offset,(byte)1) > 0;
+        public static bool ReadBit(this uint self, int offset)
+            => self.ReadBits(offset,1) > 0;
         
         /// <summary>
         /// Read a single bit from the specified offset.
@@ -238,8 +238,8 @@ namespace Jcd.BitManipulation
         /// <param name="self">the bits to read from</param>
         /// <param name="offset">the offset to read</param>
         /// <returns>the bit (a bool) at the specified offset </returns>
-        public static bool ReadBit(this int self, byte offset)
-            => self.ReadBits(offset,(byte)1) > 0;
+        public static bool ReadBit(this int self, int offset)
+            => self.ReadBits(offset,1) > 0;
         
         /// <summary>
         /// Read a single bit from the specified offset.
@@ -247,8 +247,8 @@ namespace Jcd.BitManipulation
         /// <param name="self">the bits to read from</param>
         /// <param name="offset">the offset to read</param>
         /// <returns>the bit (a bool) at the specified offset </returns>
-        public static bool ReadBit(this ushort self, byte offset)
-            => self.ReadBits(offset,(byte)1) > 0;
+        public static bool ReadBit(this ushort self, int offset)
+            => self.ReadBits(offset,1) > 0;
 
         /// <summary>
         /// Read a single bit from the specified offset.
@@ -256,8 +256,8 @@ namespace Jcd.BitManipulation
         /// <param name="self">the bits to read from</param>
         /// <param name="offset">the offset to read</param>
         /// <returns>the bit (a bool) at the specified offset </returns>
-        public static bool ReadBit(this short self, byte offset)
-            => self.ReadBits(offset,(byte)1) > 0;
+        public static bool ReadBit(this short self, int offset)
+            => self.ReadBits(offset,1) > 0;
         
         /// <summary>
         /// Read a single bit from the specified offset.
@@ -265,8 +265,8 @@ namespace Jcd.BitManipulation
         /// <param name="self">the bits to read from</param>
         /// <param name="offset">the offset to read</param>
         /// <returns>the bit (a bool) at the specified offset </returns>
-        public static bool ReadBit(this byte self, byte offset)
-            => self.ReadBits(offset,(byte)1) > 0;
+        public static bool ReadBit(this byte self, int offset)
+            => self.ReadBits(offset,1) > 0;
 
         /// <summary>
         /// Read a single bit from the specified offset.
@@ -274,8 +274,8 @@ namespace Jcd.BitManipulation
         /// <param name="self">the bits to read from</param>
         /// <param name="offset">the offset to read</param>
         /// <returns>the bit (a bool) at the specified offset </returns>
-        public static bool ReadBit(this sbyte self, byte offset)
-            => self.ReadBits(offset,(byte)1) > 0;
+        public static bool ReadBit(this sbyte self, int offset)
+            => self.ReadBits(offset,1) > 0;
         
         #endregion
