+++ Jcd.BitManipulation/SetBitsExtensions.cs
@@ -9,7 +9,7 @@ namespace Jcd.BitManipulation
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to store the value.</param>
         /// <param name="size">The number of bits to set on.</param>
-        public static void SetBits(this ref sbyte self, byte offset=0, byte size=8)
+        public static void SetBits(this ref sbyte self, int offset=0, int size=8)
         {
             self.SetBits(BitMask.CreateRange(offset, size));
         }
@@ -20,7 +20,7 @@ namespace Jcd.BitManipulation
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to store the value.</param>
         /// <param name="size">The number of bits to set on.</param>
-        public static void SetBits(this ref short self, byte offset=0, byte size=16)
+        public static void SetBits(this ref short self, int offset=0, int size=16)
         {
             self.SetBits(BitMask.CreateRange(offset, size));
         }
@@ -31,7 +31,7 @@ namespace Jcd.BitManipulation
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to store the value.</param>
         /// <param name="size">The number of bits to set on.</param>
-        public static void SetBits(this ref int self, byte offset=0, byte size=32)
+        public static void SetBits(this ref int self, int offset=0, int size=32)
         {
             self.SetBits(BitMask.CreateRange(offset, size));
         }
@@ -42,7 +42,7 @@ namespace Jcd.BitManipulation
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to store the value.</param>
         /// <param name="size">The number of bits to set on.</param>
-        public static void SetBits(this ref long self, byte offset=0, byte size=64)
+        public static void SetBits(this ref long self, int offset=0, int size=64)
         {
             self.SetBits(BitMask.CreateRange(offset, size));
         }
@@ -53,7 +53,7 @@ namespace Jcd.BitManipulation
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to store the value.</param>
         /// <param name="size">The number of bits to set on.</param>
-        public static void SetBits(this ref byte self, byte offset=0, byte size=8)
+        public static void SetBits(this ref byte self, int offset=0, int size=8)
         {
             self.SetBits(BitMask.CreateRange(offset, size));
         }
@@ -64,7 +64,7 @@ namespace Jcd.BitManipulation
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to store the value.</param>
         /// <param name="size">The number of bits to set on.</param>
-        public static void SetBits(this ref ushort self, byte offset=0, byte size=16)
+        public static void SetBits(this ref ushort self, int offset=0, int size=16)
         {
             self.SetBits(BitMask.CreateRange(offset, size));
         }
@@ -75,7 +75,7 @@ namespace Jcd.BitManipulation
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to store the value.</param>
         /// <param name="size">The number of bits to set on.</param>
-        public static void SetBits(this ref uint self, byte offset=0, byte size=32)
+        public static void SetBits(this ref uint self, int offset=0, int size=32)
         {
             self.SetBits(BitMask.CreateRange(offset, size));
         }
@@ -86,7 +86,7 @@ namespace Jcd.BitManipulation
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to store the value.</param>
         /// <param name="size">The number of bits to set on.</param>
-        public static void SetBits(this ref ulong self, byte offset=0, byte size=64)
+        public static void SetBits(this ref ulong self, int offset=0, int size=64)
         {
             self.SetBits(BitMask.CreateRange(offset, size));
         }
@@ -98,7 +98,7 @@ namespace Jcd.BitManipulation
         /// </summary>
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to set to on.</param>
-        public static void SetBit(this ref sbyte self, byte offset)
+        public static void SetBit(this ref sbyte self, int offset)
         {
             self.SetBits(BitMask.CreateSingleBit(offset));
         }
@@ -108,7 +108,7 @@ namespace Jcd.BitManipulation
         /// </summary>
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to set to on.</param>
-        public static void SetBit(this ref short self, byte offset)
+        public static void SetBit(this ref short self, int offset)
         {
             self.SetBits(BitMask.CreateSingleBit(offset));
         }
@@ -118,7 +118,7 @@ namespace Jcd.BitManipulation
         /// </summary>
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to set to on.</param>
-        public static void SetBit(this ref int self, byte offset)
+        public static void SetBit(this ref int self, int offset)
         {
             self.SetBits(BitMask.CreateSingleBit(offset));
         }
@@ -128,7 +128,7 @@ namespace Jcd.BitManipulation
         /// </summary>
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to set to on.</param>
-        public static void SetBit(this ref long self, byte offset)
+        public static void SetBit(this ref long self, int offset)
         {
             self.SetBits(BitMask.CreateSingleBit(offset));
         }
@@ -138,7 +138,7 @@ namespace Jcd.BitManipulation
         /// </summary>
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to set to on.</param>
-        public static void SetBit(this ref byte self, byte offset)
+        public static void SetBit(this ref byte self, int offset)
         {
             self.SetBits(BitMask.CreateSingleBit(offset));
         }
@@ -148,7 +148,7 @@ namespace Jcd.BitManipulation
         /// </summary>
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to set to on.</param>
-        public static void SetBit(this ref ushort self, byte offset)
+        public static void SetBit(this ref ushort self, int offset)
         {
             self.SetBits(BitMask.CreateSingleBit(offset));
         }
@@ -158,7 +158,7 @@ namespace Jcd.BitManipulation
         /// </summary>
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to set to on.</param>
-        public static void SetBit(this ref uint self, byte offset)
+        public static void SetBit(this ref uint self, int offset)
         {
             self.SetBits(BitMask.CreateSingleBit(offset));
         }
@@ -168,7 +168,7 @@ namespace Jcd.BitManipulation
         /// </summary>
         /// <param name="self">The value to be modified.</param>
         /// <param name="offset">The bit location to set to on.</param>
-        public static void SetBit(this ref ulong self, byte offset)
+        public static void SetBit(this ref ulong self, int offset)
         {
             self.SetBits(BitMask.CreateSingleBit(offset));
         }
